# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ice Truck Tracking API",
    "version": "2.0.0",
    "description": "Professional Ice Truck Tracking System API — PostgreSQL + TimescaleDB, Redis, MQTT"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Driver": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "license_number": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "User login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "minLength": 3 },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns JWT token" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "User registration",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "minLength": 3 },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Registration successful" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/drivers": {
      "get": {
        "tags": ["Drivers"],
        "summary": "Get all drivers",
        "responses": {
          "200": {
            "description": "List of drivers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Driver" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Drivers"],
        "summary": "Create a driver (requires drivers:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Driver" }
            }
          }
        },
        "responses": {
          "201": { "description": "Driver created" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
    "/drivers/{id}": {
      "get": {
        "tags": ["Drivers"],
        "summary": "Get driver by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Driver details" },
          "404": { "description": "Driver not found" }
        }
      }
    },
...
    "/feature-flags/{key}/check": {
      "get": {
        "tags": ["Feature Flags"],
        "summary": "Check if a feature flag is enabled",
        "parameters": [
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Feature flag status" },
          "404": { "description": "Feature flag not found" }
        }
      }
    },
    "/feature-flags/{key}": {
      "patch": {
        "tags": ["Feature Flags"],
        "summary": "Update feature flag (requires settings:update)",
        "parameters": [
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Feature flag updated" },
          "404": { "description": "Feature flag not found" }
        }
      }
    },
    "/feature-flags/{key}/rollout": {
      "patch": {
        "tags": ["Feature Flags"],
        "summary": "Update rollout percentage (requires settings:update)",
        "parameters": [
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rolloutPercentage": { "type": "number", "minimum": 0, "maximum": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Rollout percentage updated" },
          "404": { "description": "Feature flag not found" }
        }
      }
    },
    "/mobile/register": {
      "post": {
        "tags": ["Mobile"],
        "summary": "Register a mobile device",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["deviceId"],
                "properties": {
                  "deviceId": { "type": "string" },
                  "deviceInfo": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Device registered" }
        }
      }
    },
    "/mobile/device/{deviceId}": {
      "get": {
        "tags": ["Mobile"],
        "summary": "Get device status",
        "parameters": [
          { "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Device status" },
          "404": { "description": "Device not found" }
        }
      }
    },
    "/mobile/location": {
      "post": {
        "tags": ["Mobile"],
        "summary": "Update device location",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["deviceId", "latitude", "longitude"],
                "properties": {
                  "deviceId": { "type": "string" },
                  "latitude": { "type": "number" },
                  "longitude": { "type": "number" },
                  "accuracy": { "type": "number" },
                  "timestamp": { "type": "string", "format": "date-time" },
                  "truckId": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Location updated" }
        }
      }
    },
...
    "/mobile/sync": {
      "post": {
        "tags": ["Mobile"],
        "summary": "Simulate offline sync (batch locations)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["deviceId", "locations"],
                "properties": {
                  "deviceId": { "type": "string" },
                  "locations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "latitude": { "type": "number" },
                        "longitude": { "type": "number" },
                        "timestamp": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Locations synced" },
          "400": { "description": "Locations must be an array" }
        }
      }
    }
  }
}